fixing SentinelDatabases fixing some type errors
